requirements.txt
pyserial
numpy
pandas
matplotlib
python-dotenv
streamlit
requests
opencv-python
